Added util test; removed pkg from sonar
