+++ internal/transformer/transformer_test.go
@@ -2,6 +2,7 @@
 SPDX-FileCopyrightText: 2023 SAP SE or an SAP affiliate company and cap-operator contributors
 SPDX-License-Identifier: Apache-2.0
 */
+
 package transformer
 
 import (
