+++ sonar-project.properties
@@ -7,7 +7,7 @@ sonar.organization=sap-1
 # Go!
 sonar.language=go
 # Path is relative to the sonar-project.properties file. Replace "\" by "/" on Windows.
-sonar.sources=pkg,internal
+sonar.sources=internal
 sonar.exclusions=**/*_test.go
 sonar.tests=.
 sonar.test.inclusions=**/*_test.go
